implementacion de select de ventas

+++ DAO.Implementacion/Session.cs
@@ -8,7 +8,7 @@ namespace DAO.Implementacion
 {
     public static class  Session
     {
-        public static int IdSession;
+        public static short IdSession;
         public static string SessionUserName; 
         public static string SessionRole;       
     }
